ci: avoid running test workflow twice for branch-PRs
When pushing commits to a non-master branch and creating a
pull request from it, the CI pipeline runs twice: once
for the push, and once for the pull request. This conditional
inhibits the second run.

This is (shamelessly) copied from the Container Image workflow
in ffddorf/unms-exporter, all credits to @mraerino :)

See: https://github.com/ffddorf/unms-exporter/blob/main/.github/workflows/build-publish.yml