tests: start test server on port 0
This avoids a conflict with locally running instance ("port
already in use"). Port 0 basically allows the OS to chose a
free port number.

This however requires that we need to lookup and remember the
actual port. I've selected the (*service).start() method, and
a private field on the service struct for it - the latter is
reachable from the test and allows constructing the correct
query URL.