refactor: use ULID as Request ID and re-use as workdir name
Using the same ID allows associating the request with its (generated)
files more easily, although this is only useful with --keep-jobs=always
or --keep-jobs=on-failure.

To decrease the chance of ID collisions, we're switching from manually
generating 48 bits in base64 encoding to ULID [1], which boast a 48 bit
timestamp with ms precision and 80 bits of randomness.

[1]: https://github.com/oklog/ulid